+++ lib/widgets/viewer/info/embedded/notifications.dart
@@ -1,19 +1,21 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-enum EmbeddedDataSource { googleDevice, motionPhotoVideo, videoCover, xmp }
+enum EmbeddedDataSource { googleDevice, motionPhotoVideo, mpf, videoCover, xmp }
 
 @immutable
 class OpenEmbeddedDataNotification extends Notification {
   final EmbeddedDataSource source;
   final List<dynamic>? props;
   final String? mimeType, dataUri;
+  final int? mpfId;
 
   const OpenEmbeddedDataNotification._private({
     required this.source,
     this.props,
     this.mimeType,
     this.dataUri,
+    this.mpfId,
   });
 
   factory OpenEmbeddedDataNotification.googleDevice({
@@ -26,6 +28,11 @@ class OpenEmbeddedDataNotification extends Notification {
 
   factory OpenEmbeddedDataNotification.motionPhotoVideo() => const OpenEmbeddedDataNotification._private(
         source: EmbeddedDataSource.motionPhotoVideo,
+      );
+
+  factory OpenEmbeddedDataNotification.mpf(int id) => OpenEmbeddedDataNotification._private(
+        source: EmbeddedDataSource.mpf,
+        mpfId: id,
       );
 
   factory OpenEmbeddedDataNotification.videoCover() => const OpenEmbeddedDataNotification._private(
@@ -43,5 +50,5 @@ class OpenEmbeddedDataNotification extends Notification {
       );
 
   @override
-  String toString() => '$runtimeType#${shortHash(this)}{source=$source, props=$props, mimeType=$mimeType, dataUri=$dataUri}';
+  String toString() => '$runtimeType#${shortHash(this)}{source=$source, props=$props, mimeType=$mimeType, dataUri=$dataUri, index=$mpfId}';
 }